Fixed 2 bugs with loading tabs
1. When you loaded a new user without any lists, a New List tab appears
at the top of the page, but you couldn’t create the list using that.  I
fixed it by having the Tab component create the list if the listId is
0.  Not an ideal solution, but it works for now.

2. After I fixed that, when I loaded an existing user the selected list
was always that “new list” tab.  I realized when signing in I wasn’t
setting the selected list to the first list for that user anymore.  So
I fixed that as well.